+++ packages/cli/src/commands/service/dev.ts
@@ -6,14 +6,18 @@ import API from '@mesg/api'
 import chalk from 'chalk'
 import { decode } from '@mesg/api/lib/util/encoder'
 import { parseLog } from '../../utils/docker'
-import * as Tasks from '../../tasks'
+import * as Environment from '../../tasks/environment'
+import * as Service from '../../tasks/service'
+import * as Runner from '../../tasks/runner'
+import * as Instance from '../../tasks/instance'
 import version from '../../version'
 
 const ipfsClient = require('ipfs-http-client')
 
-type Context = Tasks.IStartEnvironment | Tasks.ICompileService | Tasks.ICreateService | Tasks.IStartService | Tasks.IServiceLogs
+type Context = Environment.IStop | Service.ICompile | Service.ICreate | Runner.ICreate | Runner.ILogs | Instance.IEventLogs | Runner.IResultLogs
+type ContextEnd = Runner.ILogsStop | Instance.IEventLogsStop | Runner.IResultLogsStop | Environment.IStop | Environment.IClearConfig
 
-export default class Service extends Command {
+export default class Dev extends Command {
   static description = 'Run a service in a local development environment'
 
   static flags = {
@@ -36,14 +40,21 @@ export default class Service extends Command {
   }]
 
   async run() {
-    const { args, flags } = this.parse(Service)
+    const { args, flags } = this.parse(Dev)
 
     const tasks = new Listr<Context>([
-      Tasks.startEnvironment,
-      Tasks.compileService,
-      Tasks.createService,
-      Tasks.startService,
-      Tasks.serviceLogs,
+      Environment.start,
+      Service.compile,
+      Service.create,
+      Runner.create,
+      {
+        title: 'Start service\'s logs',
+        task: () => new Listr<Runner.ILogs | Instance.IEventLogs | Runner.IResultLogs>([
+          Runner.logs,
+          Instance.eventLogs,
+          Runner.resultLogs,
+        ])
+      }
     ])
     const result = await tasks.run({
       configDir: flags.configDir,
@@ -58,9 +69,9 @@ export default class Service extends Command {
       lcd: new LCD('http://localhost:1317'),
     })
 
-    const runnerLogs = (result as Tasks.IRunnerLogs).runnerLogs || []
-    const eventLogs = (result as Tasks.IInstanceEventLogs).eventLogs
-    const resultLogs = (result as Tasks.IRunnerResultLogs).resultLogs
+    const runnerLogs = (result as Runner.ILogs).runnerLogs || []
+    const eventLogs = (result as Instance.IEventLogs).eventLogs
+    const resultLogs = (result as Runner.IResultLogs).resultLogs
 
     for (const log of runnerLogs) {
       log
@@ -79,12 +90,19 @@ export default class Service extends Command {
       .on('error', error => { this.warn('Result stream error: ' + error.message) })
 
     process.once('SIGINT', async () => {
-      await new Listr<Tasks.IServiceLogsStop | Tasks.IStopEnvironment | Tasks.IClearConfig>([
-        Tasks.serviceLogsStop,
-        Tasks.stopEnvironment,
-        Tasks.clearConfig
+      await new Listr<ContextEnd>([
+        {
+          title: 'Stop service\'s logs',
+          task: () => new Listr<Runner.ILogsStop | Instance.IEventLogsStop | Runner.IResultLogsStop>([
+            Runner.logsStop,
+            Instance.eventLogsStop,
+            Runner.resultLogsStop
+          ])
+        },
+        Environment.stop,
+        Environment.clearConfig
       ]).run({
-        configDir: (result as Tasks.ICreateConfig).configDir,
+        configDir: (result as Environment.ICreateConfig).configDir,
         eventLogs: eventLogs,
         resultLogs: resultLogs,
         runnerLogs: runnerLogs,
